python_microservice/requirements.txt
opencv-python
requests
Pillow
numpy
deepface
flask
pandas
tf-keras
